Added redirects to prevent seeing other teachers lessons and subjects

+++ uepel/app/Http/Controllers/LessonController.php
@@ -6,7 +6,9 @@ use App\Lesson;
 use App\Degree;
 use App\Http\Controllers\DegreeController;
 use App\Presence;
+use App\Subject;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class LessonController extends Controller
 {
@@ -14,18 +16,26 @@ class LessonController extends Controller
     {
 
         $subject = Teacher::findOrFail($subject);
+        if( $subject->teacher_id != Auth::user()->id )
+            return redirect('/teachers/home');
         return view('lessons.index' , ['teacher' => $subject]);
     }
 
     public function show($index)
     {
         $lesson = Lesson::findOrFail($index);
+        $subject = Subject::findOrFail($lesson->subject_id);
+        if( $subject->teacher_id != Auth::user()->id )
+            return redirect('/teachers/home');
         return view('lessons.show',  ['lesson' => $lesson]);
     }
 
     public function create($index)
     {
-
+        $lesson = Subject::findOrFail($index);
+        $subject = Subject::findOrFail($lesson->subject_id);
+        if( $subject->teacher_id != Auth::user()->id )
+            return redirect('/teachers/home');
         return view('lessons.create',['index'=>$index]);
     }
 
@@ -67,6 +77,11 @@ class LessonController extends Controller
     public function update($index)
     {
         $lesson = Lesson::findOrFail($index);
+
+        $subject = Subject::findOrFail($lesson->subject_id);
+        if( $subject->teacher_id != Auth::user()->id )
+            return redirect('/teachers/home');
+
         $data =request()->validate([
             'description' => '',
             'title' => 'required'
@@ -86,6 +101,11 @@ class LessonController extends Controller
     {
 
         $lesson = Lesson::findOrFail($index);
+
+        $subject = Subject::findOrFail($lesson->subject_id);
+        if( $subject->teacher_id != Auth::user()->id )
+            return redirect('/teachers/home');
+
         $this->deletePresenceDegree($index);
         $index = $lesson->subject->id;
 
@@ -98,6 +118,10 @@ class LessonController extends Controller
     public function deletePresenceDegree($index)
     {
         $lesson = Lesson::findOrFail($index);
+
+        $subject = Subject::findOrFail($lesson->subject_id);
+        if( $subject->teacher_id != Auth::user()->id )
+            return redirect('/teachers/home');
 
         $degree = Degree::all();
         foreach ($degree as $degrees){
